adding Domain layer and tests. most methods are pass through, but some like creating a review and deleting a review depend on userinfo. Added a userinfo object and a provider that will be injected into the repository Modified the ReviewDataManager so it will only delete reviews created by the user that is deleting.

+++ src/Data/RestaurantReviews.Data/ReviewDataManager.cs
@@ -9,7 +9,7 @@ namespace RestaurantReviews.Data
     public interface IReviewDataManager
     {
         Task<Review> CreateReviewAsync(Review review);
-        Task DeleteReviewAsync(int id);
+        Task DeleteReviewAsync(int id, int userId);
         Task<Review> GetReviewAsync(int id);
         Task<IEnumerable<Review>> GetReviewsAsync(int page, int pagesize, DbFilter<Review> filter);
     }
@@ -68,13 +68,13 @@ namespace RestaurantReviews.Data
             }
         }
 
-        public async Task DeleteReviewAsync(int id)
+        public async Task DeleteReviewAsync(int id, int userId)
         {
             using (var db = GetConnection())
             {
-                string query = @"DELETE from [dbo].[Review] where id = @id";
+                string query = @"DELETE from [dbo].[Review] where id = @id and user_id = @userId";
 
-                var result = await db.QueryAsync<Review>(query, new { id });
+                var result = await db.QueryAsync<Review>(query, new { id, userId});
             }
         }
